@@ -3,7 +3,8 @@
   "scripts": {
     "check": "yarn clean && yarn lint && yarn prepack:all && yarn test",
     "lint:fix": "yarn prettier:fix && yarn run eslint --fix",
-    "test": "lerna run --concurrency 1 test",
+    "pretest": "./scripts/pretest",
+    "test": "jest -i",
     "prepack:all": "scripts/prepack-all",
     "watch": "yarn workspace postgraphile prebuild && tsc -b --watch",
     "clean": "rm -Rf packages/*/dist packages/*/tsconfig.tsbuildinfo",
@@ -17,32 +18,34 @@
     "prettier:check": "yarn prettier:all --list-different"
   },
   "devDependencies": {
-    "@babel/cli": "^7.0.0",
+    "@babel/cli": "^7.8.4",
     "@babel/core": "^7.0.0",
     "@babel/plugin-syntax-object-rest-spread": "^7.0.0",
-    "@babel/plugin-transform-modules-commonjs": "^7.0.0",
-    "@babel/plugin-transform-runtime": "^7.0.0",
-    "@babel/preset-env": "^7.0.0",
-    "@typescript-eslint/eslint-plugin": "^2.3.0",
-    "@typescript-eslint/parser": "^2.3.0",
-    "@typescript-eslint/typescript-estree": "^2.3.0",
-    "babel-eslint": "^10.0.1",
-    "babel-jest": "^24.9.0",
+    "@babel/plugin-transform-modules-commonjs": "^7.9.0",
+    "@babel/plugin-transform-runtime": "^7.9.0",
+    "@babel/preset-env": "^7.9.5",
+    "@babel/preset-typescript": "^7.9.0",
+    "@typescript-eslint/eslint-plugin": "^2.29.0",
+    "@typescript-eslint/parser": "^2.29.0",
+    "@typescript-eslint/typescript-estree": "^2.29.0",
+    "babel-eslint": "^10.1.0",
+    "babel-jest": "^25.4.0",
     "concurrently": "^4.1.0",
-    "conventional-changelog-cli": "^2.0.27",
+    "conventional-changelog-cli": "^2.0.31",
     "eslint": "^6.4.0",
-    "eslint-config-prettier": "^6.10.0",
+    "eslint-config-prettier": "^6.10.1",
     "eslint-plugin-graphql": "^3.1.1",
-    "eslint-plugin-jest": "^23.7.0",
-    "eslint-plugin-tsdoc": "^0.2.1",
-    "eslint_d": "^8.0.0",
+    "eslint-plugin-jest": "^23.8.2",
+    "eslint-plugin-tsdoc": "^0.2.4",
+    "eslint_d": "^8.1.1",
     "graphql": ">=0.9 <0.14 || ^14.0.2",
-    "jest": "^24.8.0",
+    "jest": "^25.4.0",
     "lerna": "^3.18.4",
     "pg": "^7.10.0",
     "postgraphile": "4.6.0",
     "prettier": "^1.19.1",
-    "typescript": "^3.7.2"
+    "ts-jest": "^25.4.0",
+    "typescript": "^3.9.0-beta"
   },
   "workspaces": [
     "packages/*"
