@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 /* eslint-disable no-console,curly */
 import subscribeToLogicalDecoding, { AnnounceCallback } from "./index";
-import * as WebSocket from "ws";
+import WebSocket from "ws";
 
 const CONNECTION_STRING = process.env.LD_DATABASE_URL;
 const TABLE_PATTERN = process.env.LD_TABLE_PATTERN || "*.*";
