@@ -1,6 +1,6 @@
 /* eslint-disable no-console */
 import { Plugin } from "postgraphile-core";
-import * as WebSocket from "ws";
+import WebSocket from "ws";
 import subscribeToLogicalDecoding, {
   Announcement,
   LDSubscription,
